+++ R/attributes.R
@@ -8,47 +8,6 @@ dist2time <- function(dist, speed){ # TODO : need to find default value
 }
 
 
-#' minute_to_time
-#'
-#' Transform minute variable into a character string in hour format.
-#'
-#' @param time positive numeric value in minute
-#' @param sec add the sec part of the string. 
-#' @param sep ':' by default, choice between ':' and 'h'. Only affect only the
-#' first separator character
-#' @param day TRUE by default, allow to set time in a day period (betwwen 00h00 
-#' and 23:59)
-#' 
-#' @return character string
-#' 
-#' @examples 
-#' minute_to_time(130.5, sec = TRUE)
-#' minute_to_time(130.5, sec = FALSE)
-#' minute_to_time(130.5, sec = TRUE, sep = 'h')
-#' minute_to_time(130.5, sec = FALSE, sep = 'h')
-#' minute_to_time(1440, sec = FALSE, sep = 'h')
-#' minute_to_time(1664, sec = FALSE, sep = 'h')
-#'
-#' @author Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
-#'
-#' @export
-minute_to_time <- function(time, sec = TRUE, sep = c(':', 'h'), day = TRUE){
-  sep <- match.arg(sep)
-  
-  while(any(time >= 1440) & day){
-    time[time >= 1440 ] <- time[time >= 1440] - 1440
-  }
-  
-  if(sec){
-    res = sprintf("%02.0f%s%02.0f:%02.0f", time %/% 60, sep, time %% 60, 
-                  (time %% 60 %% 1) * 60 )
-  } else {
-    res = sprintf("%02.0f%s%02.0f", time %/% 60, sep, time %% 60)
-  }
-  return(res)
-}
-
-
 #' time at depth
 #'
 #' Find the time for a given depth and dive.