+++ NEWS.md
@@ -1,5 +1,13 @@
 # DiveR 0.4.0
 
+## Major improvements
+
+* Refactorisation of the dive class and all functions to cope with curve dive (input with multiples depth and time values.)
+
+* Test all of the dive class functions (except for graphic ones...)
+
+* Dive Planification Vignette.
+
 ## Historic Modifications
 
 * `max_depth_t` has been renamed to `max_depth_time`
@@ -12,9 +20,13 @@
 
 * Most summary are pretty now.
 
+* `ndive` will now call function depending on the second dive desaturation model.
+
 ## Bugs fixed
 
 * dtime work now on curve dives ( #1 )
+
+* Consecutive dives was adding interval time to total time, it has been corrected. ( #10 )
 
 * Correction of additionnal time when using successive dives. Was adding between 4 and 1 minute more than expected when using tables. ( #11 )